pyproject.toml
[project]
name = "ex-libris"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "djangorestframework>=3.16.1",
]
